Split teamController: extract members controller
